Moving AdjustTrustForm into the AddressModal folder
